move related api call to endpoints file, move global types to index.d
