+++ web/client/src/library/components/folderTree/FolderTree.tsx
@@ -1,4 +1,4 @@
-import { type Directory, type File } from '../../../api';
+import { type Directory, type File } from '../../../api/endpoints';
 import { FolderOpenIcon, DocumentIcon, FolderPlusIcon, DocumentPlusIcon, XCircleIcon } from '@heroicons/react/24/solid'
 import { FolderIcon, DocumentIcon as DocumentIconOutline } from '@heroicons/react/24/outline'
 import { ChevronRightIcon, ChevronDownIcon } from '@heroicons/react/20/solid'
