+++ web/client/src/types/index.d.ts
@@ -1,0 +1,3 @@
+type Size = typeof EnumSize[keyof typeof EnumSize]
+type Subset<T, S extends T> = S
+type Path = string
